EWPP-0000: Fixing broken page feedback form test.
