+++ modules/oe_webtools_page_feedback/tests/src/FunctionalJavascript/PageFeedbackAdminFormTest.php
@@ -8,9 +8,9 @@ use Drupal\FunctionalJavascriptTests\WebDriverTestBase;
 use Drupal\language\Entity\ConfigurableLanguage;
 
 /**
- * Tests the Page Feedback Form webtools widget.
+ * Tests the Page Feedback Form webtools widget configuration.
  */
-class PageFeedbackFormTest extends WebDriverTestBase {
+class PageFeedbackAdminFormTest extends WebDriverTestBase {
 
   /**
    * {@inheritdoc}
@@ -40,7 +40,7 @@ class PageFeedbackFormTest extends WebDriverTestBase {
   }
 
   /**
-   * Tests Page Feedback Form configuration and block rendering.
+   * Tests Page Feedback Form configuration.
    */
   public function testPageFeedbackForm(): void {
     // Create a node and a user for configuring the Page Feedback Form.
@@ -71,31 +71,12 @@ class PageFeedbackFormTest extends WebDriverTestBase {
     $this->assertEquals(TRUE, $page_feedback_config->get('enabled'));
     $this->assertEquals('1234', $page_feedback_config->get('feedback_form_id'));
 
-    // Assert the block is rendered only on node pages following the interface
-    // language.
-    $this->drupalLogout();
-    $this->drupalGet('/node/1');
-    $this->assertSession()->pageTextContains('Page node');
-    $this->assertSession()->responseContains('<script type="application/json" data-process="true">{"service":"dff","id":1234,"lang":"en"}</script>');
-    $this->drupalGet('<front>');
-    $this->assertSession()->responseNotContains('"service":"dff"');
-    $this->drupalGet('/pt-pt/node/1');
-    $this->assertSession()->responseContains('<script type="application/json" data-process="true">{"service":"dff","id":1234,"lang":"pt"}</script>');
-    $page_feedback_config->set('feedback_form_id', 12345)->save();
-    $this->drupalGet('/pt-pt/node/1');
-    $this->assertSession()->responseContains('<script type="application/json" data-process="true">{"service":"dff","id":12345,"lang":"pt"}</script>');
-
-    // Disable the block and assert the block is not rendered and the cache was
-    // properly invalidated.
+    // Disable the block and check states.
     $this->drupalLogin($user);
     $this->drupalGet('/admin/config/system/oe_webtools_page_feedback');
     $page->uncheckField('Enabled');
     $this->assertFalse($this->assertSession()->elementExists('css', 'input#edit-feedback-form-id')->hasAttribute('required'));
     $page->pressButton('Save configuration');
-    $this->drupalLogout();
-    $this->drupalGet('/node/1');
-    $this->assertSession()->pageTextContains('Page node');
-    $this->assertSession()->responseNotContains('"service":"dff"');
   }
 
 }
